only elevate priviledges after passing all checks
rootdo would setuid and setgid to root at the start of the program. If the required checks were not passes, the program would just exit. Now it only escalates priviledges at the last possible step to prevent any exploits. rdo.noconfig.c is not affected because of how simple it is